requirements.txt
asgiref==3.2.10
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
Django==3.1.1
django-common-helpers==0.9.2
django-cron==0.5.1
django-filter==2.4.0
djangorestframework==3.12.1
idna==2.10
psycopg2==2.8.6
pycodestyle==2.6.0
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
